requirements.txt
Flask==3.0.3
gunicorn==22.0.0
razorpay==1.4.2
setuptools>=70.0.0
python-dotenv==1.0.1
